Fix FinalizeSpawnEvent for trial spawners
Closes #783
