+++ patches/net/minecraft/world/level/BaseSpawner.java.patch
@@ -1,6 +1,15 @@
 --- a/net/minecraft/world/level/BaseSpawner.java
 +++ b/net/minecraft/world/level/BaseSpawner.java
-@@ -151,14 +_,14 @@
+@@ -26,7 +_,7 @@
+ import net.minecraft.world.phys.AABB;
+ import org.slf4j.Logger;
+ 
+-public abstract class BaseSpawner {
++public abstract class BaseSpawner implements net.neoforged.neoforge.common.extensions.IOwnedSpawner {
+     public static final String SPAWN_DATA_TAG = "SpawnData";
+     private static final Logger LOGGER = LogUtils.getLogger();
+     private static final int EVENT_SPAWN = 1;
+@@ -151,15 +_,14 @@
  
                          entity.moveTo(entity.getX(), entity.getY(), entity.getZ(), randomsource.nextFloat() * 360.0F, 0.0F);
                          if (entity instanceof Mob mob) {
@@ -10,26 +19,26 @@
                                  continue;
                              }
  
--                            boolean flag1 = spawndata.getEntityToSpawn().size() == 1 && spawndata.getEntityToSpawn().contains("id", 8);
+                             boolean flag1 = spawndata.getEntityToSpawn().size() == 1 && spawndata.getEntityToSpawn().contains("id", 8);
 -                            if (flag1) {
 -                                ((Mob)entity).finalizeSpawn(p_151312_, p_151312_.getCurrentDifficultyAt(entity.blockPosition()), MobSpawnType.SPAWNER, null);
-+                            // Forge: Patch in FinalizeSpawn for spawners so it may be fired unconditionally, instead of only when vanilla normally would trigger it.
-+                            var event = net.neoforged.neoforge.event.EventHooks.onFinalizeSpawnSpawner(mob, p_151312_, p_151312_.getCurrentDifficultyAt(entity.blockPosition()), null, this);
-+                            if (event != null && spawndata.getEntityToSpawn().size() == 1 && spawndata.getEntityToSpawn().contains("id", 8)) {
-+                                ((Mob)entity).finalizeSpawn(p_151312_, event.getDifficulty(), event.getSpawnType(), event.getSpawnData());
-                             }
+-                            }
++                            // Neo: Patch in FinalizeSpawn for spawners so it may be fired unconditionally, instead of only when vanilla would normally call it.
++                            // The local flag1 is the conditions under which the spawner will normally call Mob#finalizeSpawn.
++                            net.neoforged.neoforge.event.EventHooks.finalizeMobSpawnSpawner(mob, p_151312_, p_151312_.getCurrentDifficultyAt(entity.blockPosition()), MobSpawnType.SPAWNER, null, this, flag1);
  
                              spawndata.getEquipment().ifPresent(mob::equip);
-@@ -309,4 +_,12 @@
+                         }
+@@ -308,5 +_,12 @@
+ 
      public double getoSpin() {
          return this.oSpin;
-     }
-+
-+    @Nullable
-+    public Entity getSpawnerEntity() {
-+        return null;
 +    }
 +
-+    @Nullable
-+    public net.minecraft.world.level.block.entity.BlockEntity getSpawnerBlockEntity(){ return null; }
++    @Override
++    @org.jetbrains.annotations.Nullable
++    public com.mojang.datafixers.util.Either<net.minecraft.world.level.block.entity.BlockEntity, Entity> getOwner() {
++        // The vanilla anonymous classes have proper overrides, but we return null here for compatibility.
++        return null;
+     }
  }
